Move sel_points and intersect line related methods to base class
